feat(sources): split the github reader and honor branch/path filters
Split github.rs into github/{types,git,api}.rs so the orchestrator stays under
the 500-line ceiling. list_items_inner now reads MemorySourceEntry branch/paths:
git log narrows to the configured ref instead of --all and applies pathspecs,
and the REST fallback sends sha and path query params. gh_available is async
with the probe cached in a tokio OnceCell, and tests moved to the sibling
github_tests.rs.
